Split debug logic into separate file
