+++ src/dataparser/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(dataparser
         csv_parser.c
+        parse_debug.c
 )
 
 target_include_directories(dataparser PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
