+++ src/dataparser/parse_debug.h
@@ -1,0 +1,7 @@
+#include "csv_parser.h"
+
+/// @breif Used for debug printing when parsing lists of nodes and edges
+/// @param edges - struct-array of edge_t type
+void parsing_debug_print(edge_t *edges);
+/// @breif Used for debugging the parsing of lists of nodes and edges
+void parsing_debug();
